pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "contena"
version = "0.1.0"
description = "Containerized MVP for system trading validation pipelines"
readme = "README.md"
requires-python = ">=3.11"
dependencies = ["PyYAML>=6.0"]

[project.scripts]
contena = "contena.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
